+++ Magic/CHANGELOG.md
@@ -41,6 +41,7 @@
  - Add `max_spells` property to classes and wands, can be used to limit the number of spells a player may learn
  - Add spawner properties: `check_radius`, `vertical_check_radius` and `check_floor`, for controlling
    how spawners check for free space. May be needed if spawning ghasts, spiders or other large mobs.
+ - Add `block` and `stop_block` triggers, for blocking with a shield
 
 ## Fixes